+++ PlatonicRelationships/Framework/EventEditor.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
+using StardewValley.GameData;
 
 namespace PlatonicRelationships.Framework;
 
@@ -16,6 +17,9 @@ internal class EventEditor
     /// <inheritdoc cref="DataModel.RomanticEvents" />
     private readonly Dictionary<IAssetName, Dictionary<string, string>> RomanticEvents;
 
+    /// <inheritdoc cref="DataModel.RomanticTriggerActions" />
+    private readonly Dictionary<string, string> RomanticTriggerActions;
+
 
     /*********
     ** Public methods
@@ -26,15 +30,17 @@ internal class EventEditor
     public EventEditor(DataModel data, Func<string, IAssetName> parseAssetName)
     {
         this.RomanticEvents = this.ReadRomanticEvents(data, parseAssetName);
+        this.RomanticTriggerActions = data.RomanticTriggerActions;
     }
 
     /// <inheritdoc cref="IContentEvents.AssetRequested" />
     public void OnAssetRequested(AssetRequestedEventArgs e)
     {
-        // TODO: add kill contexts for platonic versions of event ("/k <event id>")
-
+        // events
         if (this.RomanticEvents.TryGetValue(e.NameWithoutLocale, out Dictionary<string, string> events))
         {
+            // TODO: add kill contexts for platonic versions of event ("/k <event id>")
+
             e.Edit(asset =>
             {
                 var data = asset.AsDictionary<string, string>().Data;
@@ -42,7 +48,22 @@ internal class EventEditor
                 foreach ((string eventId, string npcName) in events)
                 {
                     if (data.Remove(eventId, out string script))
-                        data[this.AddDatingRequirement(eventId, npcName)] = script;
+                        data[this.EditEventId(eventId, npcName)] = script;
+                }
+            });
+        }
+
+        // trigger actions
+        else if (e.NameWithoutLocale.IsEquivalentTo("Data/TriggerActions"))
+        {
+            e.Edit(asset =>
+            {
+                var data = (List<TriggerActionData>)asset.Data;
+
+                foreach (TriggerActionData action in data)
+                {
+                    if (action?.Id is not null && this.RomanticTriggerActions.TryGetValue(action.Id, out string npcName))
+                        action.Condition = this.EditConditions(action.Condition, npcName);
                 }
             });
         }
@@ -77,7 +98,7 @@ internal class EventEditor
     /// <param name="eventId">The event ID with preconditions to edit.</param>
     /// <param name="npcName">The NPC whom the NPC must date to see the event.</param>
     /// <returns>Returns the modified event ID.</returns>
-    private string AddDatingRequirement(string eventId, string npcName)
+    private string EditEventId(string eventId, string npcName)
     {
         // Add right after the event ID, to avoid conflicting with preconditions that immediately end evaluation like
         // SendMail.
@@ -86,5 +107,18 @@ internal class EventEditor
         return split.Length == 1
             ? $"{eventId}/D {npcName}"
             : $"{split[0]}/D {npcName}/{string.Join('/', split.Skip(1))}";
+    }
+
+    /// <summary>Add a dating requirement to a game state query field.</summary>
+    /// <param name="conditions">The game state query to edit.</param>
+    /// <param name="npcName">The NPC whom the NPC must date to see the event.</param>
+    /// <returns>Returns the modified game state query.</returns>
+    private string EditConditions(string conditions, string npcName)
+    {
+        string newQuery = $"PLAYER_NPC_RELATIONSHIP Current {npcName} Dating";
+
+        return string.IsNullOrWhiteSpace(conditions)
+            ? newQuery
+            : conditions.TrimEnd() + ", " + newQuery;
     }
 }